chore: Update package dependencies for contracts and webapp
This commit updates several package dependencies in both the contracts and webapp directories. Notable changes include upgrading the `@safe-global/safe-singleton-factory`, `@safe-global/safe-deployments`, and `@safe-global/safe-modules-deployments` in the contracts, as well as various dependencies in the webapp such as `@hookform/resolvers`, `@tanstack/react-query`, and `@vitejs/plugin-react`. These updates ensure compatibility with the latest features and improvements across the libraries.