End-to-End Encrypted Transaction Registration
This PR adds the `registerTransaction` method to the secret harbour,
which recovers a signer and emits an event containing the encrypted
transaction data and signature.

We store the block in storage and use a deterministic UID scheme in
order to make each transaction registration unique, allowing for very
targeted event queries that only span over a single block and (with
the UID) only return a single event log with the transaction data.
This makes event indexing not necessary for querying Harbour
transaction details, while reducing storage to only two
slots (updating the length and adding an entry to the block index
list). Under this scheme, there are a few downsides compared to the
international harbour:

1. There is no direct mapping from a Safe transaction hash to a set of
   events. This trade-off seems acceptable because Harbour clients
   need to query all signed transactions for a specific nonce anyway,
   where other transactions with the same hash can be matched.
2. You cannot read the transaction data onchain. The trade-off seems
   acceptable because the data is encrypted anyway, so there isn't
   much you can do with it onchain to begin with. Furthermore, what
   storage there is was designed to allow laser-focused event queries
   that work with nodes that do not allow large query ranges.
3. The contract cannot verify that `safeTxStructHash` corresponds to
   the encrypted data. This adds additional verification pressure on
   clients. There are some solutions to this, which will be explored
   in the future:
   - Pushing verification to the existing validator system that exists
     with Harbour
   - Using ZKPs, although _I am currently not in favour of this_, as
     it greatly increases gas costs for transaction registration
     onchain and therefore affects scalability of the solution.

I recommend reading through the E2EE test case that illustrates how
the encrypted transaction queue will work with multiple signers.