Add Benchmarking Support
This PR adds benchmarks for the Safe secret harbour, which look
promising :smile::

```
  SafeSecretHarbour [@bench]
⛽ empty_0b            : 88980
    ✔ empty transaction (47ms)
⛽ native_transfer_0b  : 89100
    ✔ native transfer
⛽ erc20_transfer_68b  : 90720
    ✔ ERC20 transfer (68 bytes)
⛽ large_tx_data_1024b : 113839
    ✔ large transaction data
⛽ append_sig_same_tx  : 76810
    ✔ additional signature to an existing transaction
```

I also refactored benchmarking altogether - I created a new
benchmarking harness in `bench.ts`, and re-expressed the existing
benchmarks using that harness. Addionally, there is a new benchmarking
case to distinguish between native token transfers and empty
transactions (previously, we were submitting transactions to an
address with no value and mislabelling it as a native token transfer).

Some minor types were reworked to work around the fact that hardhat
`SignerWithAddress`s are not compatible with ethers `Signer`s (because
of 7702 stuff, which isn't supported by Hardhat at the moment) and
just pick the signer functionality we need for our utility functions.