Optimize Storage for Block Index Array
This PR adds a new `BlockIndices` library as a specialized array for
storing block numbers. When you append to a storage dynamic array in
Solidity, it will always write to two slots:

1. The "length" slot
2. The "value" slot, in the case of values that are smaller than
   32-bytes (like `uint64`), it will pack multiple entries into a
   single value slot

The new `BlockIndices` type works in a very similar way _except_ it
also packs values in the length slots. Since we expect most arrays to
have a length of 1 or be very small (we don't expect multiple
transactions to be proposed from the same signer for the same nonce),
we can optimize for these short cases, where the bottom `uint64` of
the "length" slot holds the actual length, and the top 3 `uint64`s of
the "length" slot holds the first 3 items. With this, registering a
transaction now becomes a single `SSTORE` which shows in our
benchmarks:

**Before:**
```
  SafeSecretHarbour [@bench]
⛽ empty_0b            : 88980
    ✔ empty transaction (47ms)
⛽ native_transfer_0b  : 89100
    ✔ native transfer
⛽ erc20_transfer_68b  : 90720
    ✔ ERC20 transfer (68 bytes)
⛽ large_tx_data_1024b : 113839
    ✔ large transaction data
⛽ append_sig_same_tx  : 76810
    ✔ additional signature to an existing transaction
```

**After:**
```
  SafeSecretHarbour [@bench]
⛽ empty_0b            : 66705
    ✔ empty transaction (39ms)
⛽ native_transfer_0b  : 66813
    ✔ native transfer
⛽ erc20_transfer_68b  : 68349
    ✔ ERC20 transfer (68 bytes)
⛽ large_tx_data_1024b : 91516
    ✔ large transaction data
⛽ append_sig_same_tx  : 54523
    ✔ additional signature to an existing transaction

```

For small transactions, this is about a 25% gas improvement!