Add Encryption Key Context (#55)
This PR adds a new `context` storage field to encryption keys. This can
be used for storing a salt related to a deterministic key devivation
mechanism, while allowing for key rotation.

Specifially, the reference harbour client will implement:

1. Compute a `Date.now()` issuance date
2. Request a SIWE signature with `issuedAt` set to the issuance date
from step 1
3. Generate a public X25519 key using entropy from the signature in step
2.
4. `registerEncryptionKey(context: issuedAt, publicKey)`

This allows the key to be deterministically recomputed by reading the
context on-chain, while also making it easy to rotate keys: you would
just run through the process again, and store a new context and public
key. Clients will know that the key rotated, and if you are connected
with the same wallet to another browser, you can recompute the key.