Notarized Transaction Registration (#54)
This PR refactors the `registerTransaction` into an `enqueuTransaction`
function that has the same parameters as before. However, the new
implementation distinguishes between a few things:

1. `encryptionBlob` is not documented as being an application specific
blob that is related to a Safe transaction hash, but not specifically
the encrypted transaction data itself. It can also potentially contain
just wrapped encryption keys for a previously emitted encrypted Safe
transaction.
2. The transaction registrations are keyed per _notary_ and not per
signer. This separates the role of generating the encrypted transaction
data from the role of signing. In the case of Harbour without relaying
the `msg.sender` is the notary. This can be a proposer account that just
creates a transaction and encrypts it for each of the signers. In a
(future) relayed version, it can be the paymaster address to represent
the harbour validator set.
3. Signatures are emitted as separate events and recorded in storage
prevent signature replaying

One notable disadvantage of this structure is that you need 4 RPC
roundtrips instead of 2 to query all of the transactions and signatures
(although, you can now query all of the transactions and their data in 2
roundtrips, and separately query all of the signatures for a safe TX
hash in another 2 roundtrips):

1. `eth_call: retrieveRegistrations(chain, safe, nonce, notary)` to get
all of the registered transaction data from a trusted `notary` (for
arguments sake, lets say some proposer account; in the future, the
Harbour paymaster meaning that it was approved by the harbour validator
set).
2. A batched RPC `eth_getLogs` filtering on the event topic, the UID and
block number from the call in step 1 (for _each_ registration returned
that way)
3. A batched RPC `eth_call: retrieveSignatures([signers], safeTxHash)`
for each safe transaction hash from step 2, and each signer you care
about
4. A batched RPC `eth_getLogs` filtered on the signature event topic and
the signer for the block range returned in step 3.

Note that it is possible to split the function into multiple separate
functions: `enqueueTransaction`, `registerTransaction`, and
`registerSignature`. I do think that it will be common to submit a
transaction registration and signature _together_ so it does make sense
to have the current function anyway.