Refactor ERC-4337 Mixin
Right now, the mixin is tightly coupled with the transaction format
for the international harbour, meaning I can't really reuse it with
the secret harbour.

This PR proposes changing the mixin to be more general on both the
validation of the data to store, and storing of the data itself, in a
way that will allow the Safe secret harbour to reuse te mixin.

I opted for `executeUserOp` because it was a convinient way to pass on
user-op specific data (specifically, the address of the `paymaster`
which is used as the "notary" for the secret harbour), but I am not
attached to this implementation.

I left a small bug in there - the bundler flow in the Harbour
interface uses the `TRUSTED_PAYMASTER` contract function which is no
longer on the ABI that is being parsed (so it always fails before
making any `eth_call` anyway), but also notably missing from the
latest contract, so even if it were there, the call would fail. I am
unsure how the paymaster is supposed to be configured in the interface
ATM, so I left a TODO for now.