Interface Changes for Secret Harbour
This (draft) PR contains the complete interface changes require to
implement an encrypted queue over `SecretHarbour`.

The PR is currently **very** large and the code is a little messy. I
will be splitting it up into smaller pieces to make it reviewable, but
I wanted to share the complete code in the meantime.

Overall, the implementation was _mostly_ straightforward as I was able
to port the logic from the secret harbour contract tests to the UI.
Most of the interesting new stuff is around session management.

Note that for this initial implementation, we cut some corners with
the encryption (rest assured, in terms of efficiency in the size of
the encrypted blob and UI features, but **NOT** security):

1. We generate a random x25519 key per recipient (this is inherited
   from `jose` library that we use). Instead, we could be more optimal
   and do ECDH with the registered signer keys:
   * The proposer will no longer need to sign for themselves as a
     recipient in order to reconstruct the transaction in the
	 interface. They would be able to decode using ECDH with any of
	 the recipient's encrypted keys.
   * The public key for the randomly generated x25519 keys do not need
     be included in the blob.
2. We do not compress the JWE data that we put on chain like we do in
   the E2E contract tests.
3. We do not automatically detect when we should resubmit an encrypted
   blob for a Safe transaction in case of key rotation, and expect the
   user to manually re-queue the transaction with the updated
   encryption keys.
   * As a corollary to this, we resubmit a whole new JWE that
     re-encrypts everything (even if it isn't needed, because some
     keys are still active and can use the previous encryption blob.

I didn't want to spend time optimizing for efficiency, as there is a
real possibility that we will use blobs for submitting transaction
data onchain, in which case we have 128kB of data anyway.