Use Chain ID as EIP-712 Salt For Harbour
We added a mechanism to relay some harbour transactions on behalf of a
user (specifically, the registration of encryption keys). This causes
some UX friction with how it is current implemented, as you must
request the wallet change to a new chain in order to sign a permission
to upload an encryption key.

I decided to change the `salt` field to hold the chain ID instead of
the expected chain for a couple of reasons:

- In the case of the encryption key, while it is being registed on
  `chainId`, it is actually meant for use on **other** chains
- This allows Wallet EIP-712 signatures without a chain ID which can
  be done when the wallet is connected to any network. In particular
  for Harbour this is nice, since you are connected to the network
  where your Safe is, and don't want to change networks just to
  register a new encryption key.

I don't think there are security downsides to this: the encryption key
signature is not more or less replayable than before. Additionally,
the encryption key signature is not meant for a specific chain anyway,
so you can't really "fish" an encryption key registration on the wrong
chain (which is what this field helps protect).