Add coverage (#75)
This PR adds a code coverage workflow for the webapp.
- Closes #68 
- Webapp tests are run with coverage enabled.
- Coverage artifacts are commented as PR comment.
  -  Comment gets updated upon future pushes
- Workflows on `push` are restricted to `main` branch since otherwise
all steps are run twice on PRs which is unnecessary.

Other:
- 2 unnecessary test logs are removed/suppressed.

---

## CLA signature

With the submission of this Pull Request, I confirm that I have read and
agree to the terms of the [Contributor License
Agreement](https://github.com/safe-research/cla/blob/main/CLA.md).