Run All CI on `main`
This PR changes our workflows to run the complete CI on pushes to the
`main` branch. This ensures that CI doesn't break for any part of the
project whenever our default branch changes.

Since CI on the `push` branch doesn't get run for PRs, this means that
we don't have to worry about this affecting DevX, it happens in the
backgroud.