[Secret Harbour]: Add Button for Configuring Encryption
This PR adds a small button that sets the value of the `Harbour
Address` field to the defaul Secret Harbour deployment, to faciliate
the configuration of encryption.

The implementation is _not_ ideal IMO, but it was the path of least
resistance at the moment, and I hope to come back and refactor
settings configuration to make this more elegant. In particular, I
chose this simple solution instead of a "use encryption" checkbox, as
I felt like guaranteeing the integrity of the settings with a checkbox
is difficult (for example, what if I check use encryption and then
manually specify an international harbour deployment? The settings
form would have to use effects to keep the settings in sync, which is
a lot of work with how we use the React forms library).