Fix International Harbour Regression (#96)
This fixes a couple issues with the International harbour path that was
introduced with the Secret Harbour changes.

1. Visual :lock: bug where if you configure encrytion, go to the
unconfigure encryption, then return to the queue, your session keys
persist and the UI indicates that the queue is encrypted (even though it
isn't). The UI would function as if the queue were not encrypted, so it
is just a visual bug.
2. Fetching encryption keys should return `null` instead of throwing an
error. This would cause issues where if an existing session were
connected prior to disabling encryption.
3. Reseting the harbour address configuration leaves the
`harbourAddress` configured as `""` instead of undefined. This causes
issues as it is assumed that `harbourAddress: Address | undefined` in a
couple of places when in fact it is `harbourAddress: Address | "" |
undefined`. This is an annoying Ethers.js typing issue where `Address =
string` (compared to Viem's <code>Address = `0x${string}`</code>

Anyway, it looks like the Secret Harbour changes introduced a few
regressions (especially related to enabling/disabling encryption when
not in a fresh state). I tested queuing transactions with both the
International harbour (using Waku) and the Secret harbour, and going
back and forth between the two - so things should be more stable now.