[Web App] Use Hash Based Routing (#99)
This PR changes the history provider to use URI fragments (i.e. `#...`)
for router history. For example, this means that the settings page will
now be `harbour.safe.dev/#/settings`. From the
[docs](https://tanstack.com/router/latest/docs/framework/react/guide/history-types#hash-routing):

> Hash routing can be helpful if your server doesn't support rewrites to
index.html for HTTP requests (among other environments that don't have a
server).

This is an alternative to #98. The upside being that this is a simple
and supported solution to the problem, however, it does mean that we use
URL fragments for navigation and have pages at different URLs than
before.