[Web App]: Add Support for Transaction Builder Bundles
This PR adds support for parsing transaction builder JSON bundles to
the current batch.

As a note, I wasn't sure the best way to implement a "file dialog"
button, so comments are welcome. It was implemented this way in order
to reuse the existing `ActionCard` component, while keeping it a link
in order to indicate navigation will happen.