[Web App]: Add Support for Transaction Builder Bundles (#102)
This PR adds support for parsing transaction builder JSON bundles to the
current batch.

As a note, I wasn't sure the best way to implement a "file dialog"
button, so comments are welcome. It was implemented this way in order to
reuse the existing `ActionCard` component, while keeping it a link in
order to indicate navigation will happen.

### Demo

[Screencast From 2025-10-15
15-35-33.webm](https://github.com/user-attachments/assets/38e2bd04-1c58-426b-bcba-edefac57ebd9)